order_management_system_github/requirements.txt
fastapi>=0.100.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0,<1.5.0
sqlalchemy-utils>=0.41.0  # For database utilities
pydantic>=2.0.0
email-validator>=2.0.0  # For EmailStr validation
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
databases[sqlite]
python-multipart>=0.0.5
zeep>=4.1.0,<4.2.0  # For BSE SOAP API
razorpay>=1.3.0  # For payment gateway
aiosmtplib>=1.1.6  # For async SMTP
python-dotenv>=0.19.0  # For environment variables

